requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
pandas==2.1.4
pytz==2023.3
